Avoid splitting values for Indexed configs
