+++ xml/shared/src/main/scala/zio/config/xml/experimental/package.scala
@@ -1,8 +1,8 @@
-package zio.config
+package zio.config.xml
 
 import zio._
 
-package object xml {
+package object experimental {
   implicit class FromConfigSourceXml(c: ConfigProvider.type) {
     def fromYamlString(xml: String): ConfigProvider =
       XmlConfigProvider.fromXmlString(xml)
